refactor: move encodable structs
