+++ src/cargo/core/resolver/encode.rs
@@ -115,37 +115,14 @@ use super::{Resolve, ResolveVersion};
 use crate::core::{Dependency, GitReference, Package, PackageId, SourceId, Workspace};
 use crate::util::errors::CargoResult;
 use crate::util::interning::InternedString;
-use crate::util::{Graph, IntoUrl, internal};
+use crate::util::{Graph, internal};
 use anyhow::{Context as _, bail};
-use cargo_util_schemas::core::SourceKind;
-use serde::de;
+use cargo_util_schemas::resolve::{
+    EncodableDependency, EncodablePackageId, EncodableResolve, EncodableSourceId, Patch,
+};
 use serde::ser;
-use serde::{Deserialize, Serialize};
-use std::cmp::Ordering;
-use std::collections::{BTreeMap, HashMap, HashSet};
-use std::fmt;
-use std::str::FromStr;
+use std::collections::{HashMap, HashSet};
 use tracing::debug;
-use url::Url;
-
-/// The `Cargo.lock` structure.
-#[derive(Serialize, Deserialize, Debug)]
-pub struct EncodableResolve {
-    version: Option<u32>,
-    package: Option<Vec<EncodableDependency>>,
-    /// `root` is optional to allow backward compatibility.
-    root: Option<EncodableDependency>,
-    metadata: Option<Metadata>,
-    #[serde(default, skip_serializing_if = "Patch::is_empty")]
-    patch: Patch,
-}
-
-#[derive(Serialize, Deserialize, Debug, Default)]
-struct Patch {
-    unused: Vec<EncodableDependency>,
-}
-
-pub type Metadata = BTreeMap<String, String>;
 
 /// Convert a `Cargo.lock` to a Resolve.
 ///
@@ -520,181 +497,6 @@ fn build_path_deps(
             );
         visited.insert(pkg.package_id().source_id());
         build_pkg(&pkg, ws, ret, visited);
-    }
-}
-
-impl Patch {
-    fn is_empty(&self) -> bool {
-        self.unused.is_empty()
-    }
-}
-
-#[derive(Serialize, Deserialize, Debug, PartialOrd, Ord, PartialEq, Eq)]
-pub struct EncodableDependency {
-    name: String,
-    version: String,
-    source: Option<EncodableSourceId>,
-    checksum: Option<String>,
-    dependencies: Option<Vec<EncodablePackageId>>,
-    replace: Option<EncodablePackageId>,
-}
-
-#[derive(Debug, Clone)]
-pub struct EncodableSourceId {
-    /// Full string of the source
-    source_str: String,
-    /// Used for sources ordering
-    kind: SourceKind,
-    /// Used for sources ordering
-    url: Url,
-}
-
-impl EncodableSourceId {
-    fn new(source: String) -> CargoResult<Self> {
-        let source_str = source.clone();
-        let (kind, url) = source
-            .split_once('+')
-            .ok_or_else(|| anyhow::format_err!("invalid source `{}`", source_str))?;
-
-        let url = url.into_url()?;
-
-        let kind = match kind {
-            "git" => {
-                let reference = GitReference::from_query(url.query_pairs());
-                SourceKind::Git(reference)
-            }
-            "registry" => SourceKind::Registry,
-            "sparse" => SourceKind::SparseRegistry,
-            "path" => SourceKind::Path,
-            kind => anyhow::bail!("unsupported source protocol: {}", kind),
-        };
-
-        Ok(Self {
-            source_str,
-            kind,
-            url,
-        })
-    }
-
-    /// Encodes the inner [`SourceId`] as a URL.
-    fn as_url(&self) -> impl fmt::Display + '_ {
-        self.source_str.clone()
-    }
-}
-
-impl ser::Serialize for EncodableSourceId {
-    fn serialize<S>(&self, s: S) -> Result<S::Ok, S::Error>
-    where
-        S: ser::Serializer,
-    {
-        s.collect_str(&self.as_url())
-    }
-}
-
-impl<'de> de::Deserialize<'de> for EncodableSourceId {
-    fn deserialize<D>(d: D) -> Result<Self, D::Error>
-    where
-        D: de::Deserializer<'de>,
-    {
-        let s = String::deserialize(d)?;
-        Ok(EncodableSourceId::new(s).map_err(de::Error::custom)?)
-    }
-}
-
-impl std::hash::Hash for EncodableSourceId {
-    fn hash<H: std::hash::Hasher>(&self, state: &mut H) {
-        self.kind.hash(state);
-        self.url.hash(state);
-    }
-}
-
-impl std::cmp::PartialEq for EncodableSourceId {
-    fn eq(&self, other: &Self) -> bool {
-        self.url == other.url && self.kind == other.kind
-    }
-}
-
-impl std::cmp::Eq for EncodableSourceId {}
-
-impl PartialOrd for EncodableSourceId {
-    fn partial_cmp(&self, other: &EncodableSourceId) -> Option<Ordering> {
-        Some(self.cmp(other))
-    }
-}
-
-impl Ord for EncodableSourceId {
-    fn cmp(&self, other: &EncodableSourceId) -> Ordering {
-        self.kind
-            .cmp(&other.kind)
-            .then_with(|| self.url.cmp(&other.url))
-    }
-}
-
-#[derive(Debug, PartialOrd, Ord, PartialEq, Eq, Hash, Clone)]
-pub struct EncodablePackageId {
-    name: String,
-    version: Option<String>,
-    source: Option<EncodableSourceId>,
-}
-
-impl fmt::Display for EncodablePackageId {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{}", self.name)?;
-        if let Some(s) = &self.version {
-            write!(f, " {}", s)?;
-        }
-        if let Some(s) = &self.source {
-            write!(f, " ({})", s.as_url())?;
-        }
-        Ok(())
-    }
-}
-
-impl FromStr for EncodablePackageId {
-    type Err = anyhow::Error;
-
-    fn from_str(s: &str) -> CargoResult<EncodablePackageId> {
-        let mut s = s.splitn(3, ' ');
-        let name = s.next().unwrap();
-        let version = s.next();
-        let source_id = match s.next() {
-            Some(s) => {
-                if let Some(s) = s.strip_prefix('(').and_then(|s| s.strip_suffix(')')) {
-                    Some(EncodableSourceId::new(s.to_string())?)
-                } else {
-                    anyhow::bail!("invalid serialized PackageId")
-                }
-            }
-            None => None,
-        };
-
-        Ok(EncodablePackageId {
-            name: name.to_string(),
-            version: version.map(|v| v.to_string()),
-            source: source_id,
-        })
-    }
-}
-
-impl ser::Serialize for EncodablePackageId {
-    fn serialize<S>(&self, s: S) -> Result<S::Ok, S::Error>
-    where
-        S: ser::Serializer,
-    {
-        s.collect_str(self)
-    }
-}
-
-impl<'de> de::Deserialize<'de> for EncodablePackageId {
-    fn deserialize<D>(d: D) -> Result<EncodablePackageId, D::Error>
-    where
-        D: de::Deserializer<'de>,
-    {
-        String::deserialize(d).and_then(|string| {
-            string
-                .parse::<EncodablePackageId>()
-                .map_err(de::Error::custom)
-        })
     }
 }
